[percolation] Fix bug that could lead to segfaults
The cluster labelling code used to not allocate enough labels for the largest
possible value. The maximum possible amount of labels was correctly set as
lattice_size/2. However since we start labelling clusters from 2 and not zero,
some more space was needed in the labels array (or we could shift all accesses
to the labels array; however just increasing the array size is simpler at a
negligible memory cost).