[percolation] Bundled the PCG random number generator
This allows us to make sure that all results are reproducible in a platform
independant way. Before we relied on stdlib's rand and srand functions whose
implementation is platform dependant and indeed produced different results for
the same seed between GNU's glibc and macOS (BSD) libc.

As a bonus, PCG [1] should be, statistically speaking, a better random number
generator.

[1] http://www.pcg-random.org